refactor: default ui
