+++ packages/playground/src/components/mcp/tools-default-view/index.ts
@@ -1,0 +1,1 @@
+export * from "./tools-default-view";
