+++ packages/playground/src/components/mcp/tools-view.tsx
@@ -3,8 +3,7 @@ import { observer } from "mobx-react-lite";
 import { useEffect, useRef, useState } from "react";
 import { Env, type MCPToolRequest, usePixel } from "@semoss/sdk/react";
 import { Skeleton } from "@semoss/ui/next";
-import type { RoomStore } from "@/stores";
-import type { MCPTool } from "@/types";
+import type { RoomStore, ToolStore } from "@/stores";
 import { ToolsDefaultView } from "./tools-default-view";
 
 const PLATFORM_URL = import.meta.env.VITE_PLATFORM_URL
@@ -22,12 +21,7 @@ interface ToolsViewProps {
 	message: string;
 
 	/** Connected tool */
-	tool: {
-		id: string;
-		name: string;
-		parameters: Record<string, unknown>;
-		original_name: string;
-	};
+	tool: ToolStore["json"];
 
 	/** Response to the tool */
 	toolResponse?: string;
@@ -44,7 +38,6 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 		const iframeRef = useRef<HTMLIFrameElement>(null);
 		const [isLoading, setIsLoading] = useState<boolean>(true);
 		const [url, setUrl] = useState("");
-		const [selectedTool, setSelectedTool] = useState<MCPTool>(null);
 
 		/**
 		 * Library Hooks
@@ -56,45 +49,6 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 		}>(app ? `ProjectInfo(project=["${app}"]);` : "", {
 			data: {
 				project_type: "",
-			},
-		});
-
-		// Get tool JSON
-		const getMCP = usePixel<{
-			_meta: {
-				SMSS_PROJECT_NAME: string;
-				SMSS_PROJECT_ID: string;
-				SMSS_ENGINE_NAME: string;
-				SMSS_ENGINE_TYPE: string;
-				SMSS_ENGINE_ID: string;
-			};
-			tools: MCPTool[];
-		}>(`GetMCPTools(project=["${app}"]);`, {
-			data: {
-				tools: [
-					{
-						name: "",
-						description: "",
-						title: "",
-						inputSchema: {
-							properties: {},
-							type: "object",
-							required: [],
-							title: "",
-						},
-						original_name: "",
-						_meta: {
-							generated_on: "",
-						},
-					},
-				],
-				_meta: {
-					SMSS_ENGINE_ID: "",
-					SMSS_ENGINE_NAME: "",
-					SMSS_ENGINE_TYPE: "",
-					SMSS_PROJECT_ID: "",
-					SMSS_PROJECT_NAME: "",
-				},
 			},
 		});
 
@@ -117,7 +71,7 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 						name: tool?.name || "",
 						parameters: toJS(tool?.parameters || {}),
 						roomId: room.roomId,
-						original_name: selectedTool?.original_name || "",
+						original_name: tool.original_name || "",
 						tool_response: toolResponse,
 						executedParameters: executedParameters,
 					} satisfies MCPToolRequest,
@@ -129,17 +83,6 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 		/**
 		 * Effects
 		 */
-
-		// Initialize selected tool from tool info
-		useEffect(() => {
-			if (getMCP.status === "SUCCESS" && tool?.original_name) {
-				setSelectedTool(
-					getMCP.data.tools.find(
-						(a) => a.name === tool.original_name,
-					),
-				);
-			}
-		}, [getMCP, tool.original_name]);
 
 		useEffect(() => {
 			const chooseUrl = async () => {
@@ -160,7 +103,7 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 
 				setIsLoading(true);
 
-				if (!selectedTool._meta.SMSS_MCP_UI) {
+				if (!tool._meta.SMSS_MCP_UI) {
 					// Legacy, check for portals
 
 					if (getAppInfo.data.project_type === "BLOCKS") {
@@ -192,8 +135,7 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 					);
 				} else {
 					// Modern
-					const resourceURI =
-						selectedTool._meta.SMSS_MCP_UI?.resourceURI;
+					const resourceURI = tool._meta.SMSS_MCP_UI?.resourceURI;
 					if (!resourceURI) {
 						// No UI defined, show form
 						setUrl(null);
@@ -211,7 +153,7 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 			};
 
 			chooseUrl();
-		}, [app, tool, getAppInfo.status, getAppInfo.data, selectedTool]);
+		}, [app, tool, getAppInfo.status, getAppInfo.data]);
 
 		return (
 			<div className="relative flex h-full w-full flex-col items-center justify-center overflow-hidden">
@@ -225,13 +167,12 @@ export const ToolsView: React.FC<ToolsViewProps> = observer(
 						onLoad={() => handleOnLoad()}
 					/>
 				)}
-				{!url && !isLoading && selectedTool && (
+				{!url && !isLoading && tool && (
 					<ToolsDefaultView
 						room={room}
 						app={app}
 						message={message}
 						tool={tool}
-						mcp={selectedTool}
 						toolResponse={toolResponse}
 						executedParameters={executedParameters}
 					/>
