+++ packages/playground/src/types.d.ts
@@ -192,6 +192,9 @@ interface ResponseToolPixelMessage extends AbstractPixelMessage {
 		/** THIS IS A STRING, but ONLY in playground we parse as an app */
 		/** THIS IS NOT USED IF THERE IS AN INPUT_TOOL_EXEC WITH THE SAME TOOL ID */
 		arguments: Record<string, unknown>;
+
+		/**  description of tool **/
+		description: string;
 	}[];
 	modelId: string;
 	ornaments: {
